+++ src/main/java/com/booking/dto/request/VNPayRequest.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class VNPayRequest {
     private String txnRef;
-    private Long total;
+    private BigDecimal total;
 }
